moved headers to interface and update bugfix in buildfile
