+++ L1Trigger/DTPhase2Trigger/interface/MotherGrouping.h
@@ -20,9 +20,9 @@
 #include "DataFormats/L1DTTrackFinder/interface/L1MuDTChambThContainer.h"
 #include "DataFormats/L1DTTrackFinder/interface/L1MuDTChambThDigi.h"
 
-#include "L1Trigger/DTPhase2Trigger/src/muonpath.h"
-#include "L1Trigger/DTPhase2Trigger/src/analtypedefs.h"
-#include "L1Trigger/DTPhase2Trigger/src/constants.h"
+#include "L1Trigger/DTPhase2Trigger/interface/muonpath.h"
+#include "L1Trigger/DTPhase2Trigger/interface/analtypedefs.h"
+#include "L1Trigger/DTPhase2Trigger/interface/constants.h"
 
 #include "L1Trigger/DTPhase2Trigger/interface/MotherGrouping.h"
